refactor folder struncture
